api_gateway/pyproject.toml
[tool.poetry]
name = "api-gateway"
version = "0.1.0"
description = ""
authors = ["bitizen248 <van4011@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
authlib = "^1.2.0"
httpx = "^0.23.3"
oauth2client = "^4.1.3"
python-dotenv = "^0.21.1"
tortoise-orm = {version = "^0.19.3", extras = ["asyncpg"]}
aerich = "^0.7.1"



[tool.aerich]
tortoise_orm = "api_gateway.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
